Add files via upload
One of the most common marketing strategies In the Banking sector is direct marketing campaigns through phone calls, it is a form of advertising that allows organizations to communicate directly with customers to offer their services based on the client’s existing bank profile. We will convert this problem into a machine-learning classification problem and develop a predictive framework. Using those models to predict whether a client will subscribe to a term deposit or not.